Fix issues.
