+++ changelog/fragments/1775092603-add-osquerybeat-otel-receiver.yaml
@@ -1,0 +1,8 @@
+# REQUIRED
+kind: feature
+
+# REQUIRED
+summary: Add osquerybeatreceiver to run osquerybeat under the EDOT collector
+
+# REQUIRED
+component: osquerybeat
